Rename JS wrappers
