--- packages/utils/node-resolver-core/index-with-fallback.js
@@ -1,5 +1,0 @@
-try {
-  module.exports = require('./index.js');
-} catch (e) {
-  module.exports = require('./index-wasm.js');
-}
