+++ packages/utils/node-resolver-core/native-wasm-browser.js
@@ -1,23 +1,22 @@
 /* eslint-disable no-undef */
 const {Environment, napi} = require('napi-wasm');
-const fs = require('fs');
-const path = require('path');
 
 module.exports.Resolver = undefined;
 module.exports.init = async function init(input) {
   if (module.exports.Resolver == null) {
-    input = input ?? path.join(__dirname, 'parcel_resolver_node.wasm');
-    if (
-      typeof input === 'string' ||
-      (typeof Request === 'function' && input instanceof Request) ||
-      (typeof URL === 'function' && input instanceof URL)
-    ) {
-      input = fs.readFileSync(input);
-    }
-
-    const {instance} = await WebAssembly.instantiate(await input, {
+    input = input ?? new URL('parcel_resolver_node.wasm', import.meta.url);
+    const {instance} = await WebAssembly.instantiateStreaming(fetch(input), {
       env: napi,
     });
+
+    // const {instance} = await WebAssembly.instantiate(
+    //   require('fs').readFileSync(
+    //     input ?? require('path').join(__dirname, 'parcel_resolver_node.wasm'),
+    //   ),
+    //   {
+    //     env: napi,
+    //   },
+    // );
 
     for (let key in instance.exports) {
       if (key.startsWith('__napi_register__')) {
