+++ packages/utils/node-resolver-core/src/index.js
@@ -1,3 +1,3 @@
 // @flow
 export {default} from './Wrapper';
-export {Resolver as ResolverBase, init} from '../index-with-fallback';
+export {Resolver as ResolverBase, init} from '../native';
